add exit

+++ src/parsing/divider.c
@@ -14,9 +14,9 @@
 
 // chekc if there's pipe and give their position in the char *in
 
-int	check_pipe(char *in)
+size_t 	check_pipe(char *in)
 {
-	int	i;
+	size_t	i;
 	int	flag;
 
 	i = 0;
